Separate duty cycle test for easier calculation and understanding
